Bugfix: migrations renaming columns should now be working

+++ laravel/database/migrations/2025_04_03_102345_rename_bulk_upload_dataset_name_prefix_column_to_bulk_upload_dataset_name_filter.php
@@ -11,10 +11,21 @@ return new class extends Migration
 	 */
 	public function up(): void
 	{
-		Schema::table('databases', function (Blueprint $table) {
-			$table->renameColumn('bulk_upload_dataset_name_prefix', 'bulk_upload_dataset_name_filter');
-			$table->dropColumn('bulk_upload_dataset_name_suffix');
-		});
+		// if column exists, then rename it
+		if (Schema::hasColumn('databases', 'bulk_upload_dataset_name_prefix'))
+		{
+			Schema::table('databases', function (Blueprint $table) {
+				$table->renameColumn('bulk_upload_dataset_name_prefix', 'bulk_upload_dataset_name_filter');
+				$table->dropColumn('bulk_upload_dataset_name_suffix');
+			});
+		}
+		// else create it
+		else
+		{
+			Schema::table('databases', function (Blueprint $table) {
+				$table->string('bulk_upload_dataset_name_filter')->nullable();
+			});
+		}
 	}
 
 	/**
@@ -24,7 +35,6 @@ return new class extends Migration
 	{
 		Schema::table('databases', function (Blueprint $table) {
 			$table->renameColumn('bulk_upload_dataset_name_filter', 'bulk_upload_dataset_name_prefix');
-            $table->string('bulk_upload_dataset_name_suffix')->nullable();
 		});
 	}
 };
